refactor: extract llm interface core types
